Added solved problem -  Missing Number

+++ src/main/java/_268.java
@@ -7,12 +7,12 @@ public class _268 {
 
     public static class Solution {
         public int missingNumber(int[] nums) {
-            long result = 0;
+            int result = 0;
             for (int i = 0; i < nums.length; ++i) {
                 result -= nums[i];
                 result += i;
             }
-            return (int)(result + nums.length);
+            return result + nums.length;
         }
     }
 }
